+++ src/lib/constants/routes.ts
@@ -1,6 +1,3 @@
 export const PATHS = {
-  REFERENCE_GENRE_TREE: "/reference-genre-tree",
   ABOUT: "/about",
-  PROTOTYPE_ROOT: "/prototype",
-  PROTOTYPE_REFERENCE_GENRE_TREE: "/prototype/reference-genre-tree",
 };